% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Assert.differs}
\alias{Assert.differs}
\title{Method Assert.differs}
\usage{
Assert.differs(object, meth, act, exp)
}
\arguments{
\item{object}{Description.}

\item{meth}{Description.}

\item{act}{Description.}

\item{exp}{Description.}
}
\value{
result Description.
}
\description{
Description.
}
\examples{
Assert.Assert.differs()
}

