% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Assert.setResult<-}
\alias{Assert.setResult<-}
\title{Method Assert.setResult}
\usage{
Assert.setResult(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{Description.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
Assert.setResult()
}

